requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.4.3
streamlit==1.14.0
